make item and storage APIs safer

+++ src/pages/api/update-storage-full.ts
@@ -1,7 +1,7 @@
-import { Prisma } from "@prisma/client";
 import { RelationProps } from "components/Widget/RelationWidget";
-import type { NextApiRequest, NextApiResponse } from "next";
+import type { NextApiResponse } from "next";
 import prisma from "services/prisma";
+import { prismaErrHandler } from "services/prismaErrHandler";
 import { TypedRequestBody } from "services/types";
 
 export default async function handle(
@@ -18,18 +18,26 @@ export default async function handle(
     count: relation.count,
     itemId: relation.itemId,
   }));
-  const op = await prisma.storage.update({
-    where: {
-      id,
-    },
-    data: {
-      name: newName,
-      description: newDescription,
-      relations: {
-        deleteMany: {},
-        createMany: { data: relations },
+  if (newName == "")
+    return res.status(500).json("Object name can not be empty");
+  try {
+    const op = await prisma.storage.update({
+      where: {
+        id,
       },
-    },
-  });
-  return res.status(200).json(op);
+      data: {
+        name: newName,
+        description: newDescription,
+        relations: {
+          deleteMany: {},
+          createMany: { data: relations },
+        },
+      },
+    });
+    return res.status(200).json(op);
+  } catch (e) {
+    const msg = prismaErrHandler(e);
+    if (msg != "0") return res.status(500).json(msg);
+    else return res.status(500).json("Unknown error: " + e);
+  }
 }
